플레이어 무적, 넉백, 발사체 데미지 적용
Okay, now gameUI class (gameUI).I'm trying to create a UI by creating a py file). The first thing I'm going to make is a timer located at the top and center of the screen.
The timer measures the time after the start of the game and indicates the time elapsed until the set time (10 minutes first).
For example, 3:41 is expressed as 03:41.
The game ends when the time is set. The ending part is menu.Use the game_over function in py